Deleting algo.rs and init.rs
