+++ src/centralized_telescope/telescope.rs
@@ -1,7 +1,5 @@
 //! Customer facing Telescope structure
 
-use super::algorithm::{bin_hash, proof_hash, prove_index};
-use super::init::setup;
 use super::params::Params;
 use super::proof::Proof;
 use super::round::Round;
@@ -30,7 +28,7 @@ impl Telescope {
         set_size: u64,
         lower_bound: u64,
     ) -> Self {
-        let params = setup(soundness_param, completeness_param, set_size, lower_bound);
+        let params = Params::setup(soundness_param, completeness_param, set_size, lower_bound);
         Telescope {
             soundness_param,
             completeness_param,
@@ -63,7 +61,7 @@ impl Telescope {
     pub fn prove(self, prover_set: &[crate::utils::types::Element]) -> Option<Proof> {
         // Run prove_index up to max_retries times
         (0..self.params.max_retries).find_map(|retry_counter| {
-            prove_index(self.set_size, &self.params, prover_set, retry_counter).1
+            Proof::prove_index(self.set_size, &self.params, prover_set, retry_counter).1
         })
     }
 
@@ -86,7 +84,7 @@ impl Telescope {
         // For each element in the proof's sequence
         for &element in &proof.element_sequence {
             // Retrieve the bin id associated to this new element
-            let Some(bin_id) = bin_hash(self.set_size, proof.retry_counter, element) else {
+            let Some(bin_id) = Proof::bin_hash(self.set_size, proof.retry_counter, element) else {
                 return false;
             };
             // Check that the new element was chosen correctly
@@ -100,7 +98,7 @@ impl Telescope {
                 return false;
             }
         }
-        proof_hash(self.params.valid_proof_probability, &round)
+        Proof::proof_hash(self.params.valid_proof_probability, &round)
     }
 }
